Extract add values widget
